refactor: tabs komponent
- Bruk nanoid til å generere ID
- Endre styling på focus outline
- Exporter én komponent per modul
- Fiks ordelegging i dokumentasjon
- Fiks index.html så `yarn dev`fungerer igjen
