+++ packages/tabs-react/documentation/Tabs.mdx
@@ -8,11 +8,11 @@ group: knapper
 import TabsExample, { tabsExampleCode } from "./TabsExample";
 import "@fremtind/jkl-tabs/tabs.css";
 
-<Ingress>En liste med tabs kan brukes til å bytte mellom innhold på en side.</Ingress>
+<Ingress>En liste med tabs kan brukes til å bytte mellom likestilt innhold på en side.</Ingress>
 
 <ComponentExample title="Tabs" component={TabsExample} codeExample={tabsExampleCode} />
 
 ## Bruk
 
 -   Denne komponenten egner seg ikke som hovednavigasjon i en applikasjon.
--   Unngå å bruk for mange nivåer med tabs, og bruk de helst ikke rett under hverandre.
+-   Unngå helst å bruke flere nivåer av tabs, og bruk dem ikke rett under hverandre.
