+++ packages/tabs-react/documentation/index.html
@@ -1,14 +1,14 @@
 <!DOCTYPE html>
-<html style="height: 100%">
+<html>
     <head>
         <meta charset="utf-8" />
         <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-        <title>Jøkul Tabs Example</title>
         <meta name="viewport" content="width=device-width, initial-scale=1" />
+        <title>Jøkul Tabs Example</title>
     </head>
 
-    <body style="height: 100%; display: flex; align-items: center; justify-content: center">
+    <body>
         <div id="app" class="jkl"></div>
-        <script src="index.tsx"></script>
+        <script type="module" src="index.tsx"></script>
     </body>
 </html>
