+++ packages/tabs-react/package.json
@@ -1,6 +1,6 @@
 {
     "name": "@fremtind/jkl-tabs-react",
-    "version": "0.0.1",
+    "version": "1.0.0",
     "publishConfig": {
         "access": "public"
     },
@@ -34,7 +34,8 @@
     "dependencies": {
         "@babel/runtime": "^7.9.0",
         "@fremtind/jkl-core": "^8.0.0",
-        "@fremtind/jkl-tabs": "^0.0.1"
+        "@fremtind/jkl-tabs": "^0.0.1",
+        "nanoid": "^3.1.30"
     },
     "peerDependencies": {
         "@types/react": "^16.8.6 || ^17.0.0",
