+++ packages/tabs-react/src/index.ts
@@ -1,1 +1,4 @@
 export * from "./Tabs";
+export * from "./TabList";
+export * from "./Tab";
+export * from "./TabPanel";
